+++ input/RightCommand.py
@@ -1,6 +1,6 @@
-from InputCommand import InputCommand
+from input.InputCommand import InputCommand
 
-class DownCommand (InputCommand):
+class RightCommand (InputCommand):
   def __init__ (self):
     InputCommand.__init__ (self)
 
@@ -8,15 +8,12 @@ class DownCommand (InputCommand):
     page = grid.get_current_page ()
     grid_elements = page.get_grid_elements ()
     selected = page.get_selected ()
-    row_length = page.get_row_length ()
 
     # unselect previous grid element
     grid_elements[selected].toggle_selected ()
 
     # calculate new selected element
-    new_selected = selected + row_length
-    if new_selected >= len (grid_elements):
-      new_selected = selected % row_length
+    new_selected = (selected + 1) % len (grid_elements)
 
     page.set_selected (new_selected)
